Move branch v0.70 to tag

+++ test/TMV_TestBandDiv.cpp
@@ -1,10 +1,10 @@
 
 #define START 0
 
-#include "TMV_Test.h"
-#include "TMV_Test_2.h"
 #include "TMV.h"
 #include "TMV_Band.h"
+#include "TMV_Test.h"
+#include "TMV_Test_2.h"
 #include "TMV_TestBandArith.h"
 
 template <class T> 
@@ -67,7 +67,7 @@ void TestBandDiv(tmv::DivType dt)
         Assert(bi.checkDecomp(divout),"CheckDecomp");
         T eps = m.rowsize()*EPS*Norm(m)*Norm(m.inverse());
 
-        if (bi.colsize() == size_t(N)) {
+        if (bi.colsize() == N) {
             tmv::Vector<T> x1 = v1/bi;
             tmv::Vector<T> x2 = v1/m;
             if (showacc) {
@@ -77,7 +77,7 @@ void TestBandDiv(tmv::DivType dt)
             Assert(Norm(x1-x2) < eps*Norm(x1),"Band v/b");
         }
 
-        if (bi.rowsize() == size_t(N)) {
+        if (bi.rowsize() == N) {
             tmv::Vector<T> x1 = v1%bi;
             tmv::Vector<T> x2 = v1%m;
             if (showacc) {
@@ -153,7 +153,7 @@ void TestBandDiv(tmv::DivType dt)
         cv(1) += std::complex<T>(-1,5);
         cv(2) -= std::complex<T>(-1,5);
 
-        if (m.colsize() == size_t(N)) {
+        if (m.colsize() == N) {
             // test real / complex
             tmv::Vector<std::complex<T> > y1 = v1/cbi;
             tmv::Vector<std::complex<T> > y2 = v1/cm;
@@ -182,7 +182,7 @@ void TestBandDiv(tmv::DivType dt)
             Assert(Norm(y1-y2) < ceps*Norm(y1),"Band cv/cb");
         }
 
-        if (bi.rowsize() == size_t(N)) {
+        if (bi.rowsize() == N) {
             tmv::Vector<std::complex<T> > y1 = v1%cbi;
             tmv::Vector<std::complex<T> > y2 = v1%cm;
             if (showacc) {
